idk anymore....
all of the adjustments i make, aren't working. i ended up changing it all back to how it was before. It will not display with github because it says i am missing the index.html file when its all in the folder. i amd very frustrated because this weekend nothing has been going well and this website is making it worse for myself.